socket.io setup and working
a seperate file for axios i.e axiosConfig  is setup so that the app can work on any cutom base  url for api calls. was using https(not working) after that tried to use only http then it worked. In socket io setup messages was not updating then debugged and found out that on receiving the new message it was not adding to the messages array due b/c some if else condition were used.